Extract src/worktree.ts — the new-session dialog
The largest single cluster in main.ts, out whole: 932 lines covering the dialog
that picks where a session starts (repo root, existing worktree, branch to check
out, or a new worktree), the branch chooser popover, and the worktree
remove/switch/delete-branch flows. main.ts 4,462 -> 3,565.

It moves as one unit rather than as a view/controller pair, which is the
prediction PLAN made about these dialogs and it held: the markup reads the
dialog's own state — which row is armed, what has been prefetched, whether a
create is in flight — so there is no seam to split it on. Its click and key
handlers moved with it too; rows are addressed by index into wtRows, so they
never belonged to the global [data-*] dispatcher anyway. Byte-identical to
HEAD:src/main.ts apart from `export` keywords.

Five hooks, and they are all one kind of thing: this dialog *acts on panes*,
which main.ts owns — launch, closeSession, setActive, renderAll, handToTerminal.
Nothing else needed one, because dlog and toast are now importable.

Three shared things moved to earn that:

- `toast` joins `$` in ./dom. It is a DOM primitive every dialog needs, and
  keeping it in main.ts would have cost a hook per dialog.
- `SCRIM_DLGS` + `dropScrim` also to ./dom — the shared backdrop is coordination
  between overlays, not anything one of them owns.
- the engine table (`EngineDef`, `ALL_ENGINES`, `engineDef`, `availEngines`) to
  ./state beside the `termEngine` preference it describes.

Smoke lesson, worth more than the smoke: a dynamic import in the browser gets a
SECOND module instance whenever vite is serving the file under a ?t= URL, and for
a module that registers event listeners that means two sets of them. Escape
appeared to close the dialog and clear the filter at once — two layers per press.
Restarting vite so every module is unchanged-since-start makes the URLs canonical;
with one instance Escape peels exactly one layer, filter first, dialog second, as
written. Nothing was wrong with the code.
